[fix]죽음 조건 체크

+++ Assets/Script/ScriptSO/UnitSO.cs
@@ -9,7 +9,7 @@ public class UnitSO : ScriptableObject
     [Header("UnitSO")]
     public int atk;         // 공격력
     public int hp;          // 체력
-    public int unitSpeed;   // 이동 속도
+    public float unitSpeed;   // 이동 속도
     public float atkSpeed;    // 공격 속도
     public int unitScale;   // 유닛 크기
